refactor: base 패키지 생성
App,Rq 클래스를 base 패키지로 이동시킴
