+++ src/main/java/com/ll/Main.java
@@ -1,5 +1,7 @@
 package com.ll;
 
+import com.ll.base.App;
+
 public class Main {
     public static void main(String[] args) {
         App app = new App();
